feat: add executor class
